requirements.txt
asgiref==3.9.1
django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-guardian==2.4.0
pyjwt==2.10.1
tzdata==2025.2

django-cors-headers
python-dotenv
